+++ bc/settings/base.py
@@ -54,7 +54,7 @@ INSTALLED_APPS = [
     "scout_apm.django",
     "bc.documents",
     "bc.events",
-    # "bc.forms",
+    "bc.forms",
     "bc.home",
     "bc.images",
     "bc.navigation",
@@ -659,6 +659,9 @@ PATTERN_LIBRARY_TEMPLATE_DIR = os.path.join(
 
 # Google Tag Manager ID from env
 GOOGLE_TAG_MANAGER_ID = env.get("GOOGLE_TAG_MANAGER_ID")
+
+# For Yandex search indexing verification
+YANDEX_VERIFICATION_STRING = env.get("YANDEX_VERIFICATION_STRING")
 
 
 # GOV.UK Notify service