+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 5.0 (2020-03-20)
+
+Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/4.0...5.0
+
+- Enable FormPages
+- Add Yandex and Bing verification strings
+
 ## 4.0 (2020-02-24)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/3.2...4.0